+++ src/java/org/malbino/orion/controllers/DashboardController.java
@@ -8,11 +8,15 @@ import java.io.IOException;
 import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 import javax.annotation.PostConstruct;
 import javax.ejb.EJB;
 import javax.enterprise.context.SessionScoped;
 import javax.inject.Named;
 import org.malbino.orion.entities.Carrera;
+import org.malbino.orion.entities.GestionAcademica;
+import org.malbino.orion.entities.Inscrito;
+import org.malbino.orion.enums.Condicion;
 import org.malbino.orion.enums.Nivel;
 import org.malbino.orion.enums.Sexo;
 import org.malbino.orion.facades.GrupoFacade;
@@ -45,18 +49,20 @@ public class DashboardController extends AbstractController implements Serializa
     @EJB
     GrupoFacade grupoFacade;
 
-    private Long totalInscritos;
-    private Long varones;
+    private GestionAcademica seleccionGestionAcademica;
+
+    private Integer totalInscritos;
+    private Integer varones;
     private Double porcentajeVarones;
-    private Long mujeres;
+    private Integer mujeres;
     private Double porcentajeMujeres;
-    private Long efectivos;
+    private Integer efectivos;
     private Double porcentajeEfectivos;
-    private Long retirados;
+    private Integer retirados;
     private Double porcentajeRetirados;
-    private Long aprobados;
+    private Integer aprobados;
     private Double porcentajeAprobados;
-    private Long reprobados;
+    private Integer reprobados;
     private Double porcentajeReprobados;
 
     private BarChartModel hbarModel;
@@ -70,22 +76,23 @@ public class DashboardController extends AbstractController implements Serializa
 
     @PostConstruct
     public void init() {
-        totalInscritos = inscritoFacade.cantidadInscritos();
+        varones = 0;
+        porcentajeVarones = 0.0;
 
-        varones = inscritoFacade.cantidadInscritos(Sexo.MASCULINO);
-        porcentajeVarones = Redondeo.redondear_HALFUP(((varones.doubleValue() / totalInscritos.doubleValue()) * 100.0), 2);
-        mujeres = inscritoFacade.cantidadInscritos(Sexo.FEMENINO);
-        porcentajeMujeres = Redondeo.redondear_HALFUP(((mujeres.doubleValue() / totalInscritos.doubleValue()) * 100.0), 2);
+        mujeres = 0;
+        porcentajeMujeres = 0.0;
 
-        efectivos = inscritoFacade.cantidadInscritosEfectivos();
-        porcentajeEfectivos = Redondeo.redondear_HALFUP(((efectivos.doubleValue() / totalInscritos.doubleValue()) * 100.0), 2);
-        retirados = totalInscritos - efectivos;
-        porcentajeRetirados = Redondeo.redondear_HALFUP(((retirados.doubleValue() / totalInscritos.doubleValue()) * 100.0), 2);
+        retirados = 0;
+        porcentajeRetirados = 0.0;
 
-        reprobados = inscritoFacade.cantidadInscritosReprobados();
-        aprobados = totalInscritos - reprobados;
-        porcentajeAprobados = Redondeo.redondear_HALFUP(((aprobados.doubleValue() / totalInscritos.doubleValue()) * 100.0), 2);
-        porcentajeReprobados = Redondeo.redondear_HALFUP(((reprobados.doubleValue() / totalInscritos.doubleValue()) * 100.0), 2);
+        efectivos = 0;
+        porcentajeEfectivos = 0.0;
+
+        aprobados = 0;
+        porcentajeAprobados = 0.0;
+
+        reprobados = 0;
+        porcentajeReprobados = 0.0;
 
         createHorizontalBarModel();
         createPieModel();
@@ -98,31 +105,76 @@ public class DashboardController extends AbstractController implements Serializa
     }
 
     public void reinit() {
-        totalInscritos = inscritoFacade.cantidadInscritos();
+        if (seleccionGestionAcademica != null) {
+            List<Inscrito> listaInscritos = inscritoFacade.listaInscritos(seleccionGestionAcademica.getId_gestionacademica());
+            totalInscritos = listaInscritos.size();
 
-        varones = inscritoFacade.cantidadInscritos(Sexo.MASCULINO);
-        porcentajeVarones = Redondeo.redondear_HALFUP(((varones.doubleValue() / totalInscritos.doubleValue()) * 100.0), 2);
-        mujeres = inscritoFacade.cantidadInscritos(Sexo.FEMENINO);
-        porcentajeMujeres = Redondeo.redondear_HALFUP(((mujeres.doubleValue() / totalInscritos.doubleValue()) * 100.0), 2);
+            if (totalInscritos > 0) {
+                List<Inscrito> listaVarones = listaInscritos.stream().filter(i -> i.getEstudiante().getSexo().equals(Sexo.MASCULINO)).collect(Collectors.toList());
+                varones = listaVarones.size();
+                porcentajeVarones = Redondeo.redondear_HALFUP(((varones.doubleValue() / totalInscritos.doubleValue()) * 100.0), 2);
 
-        efectivos = inscritoFacade.cantidadInscritosEfectivos();
-        porcentajeEfectivos = Redondeo.redondear_HALFUP(((efectivos.doubleValue() / totalInscritos.doubleValue()) * 100.0), 2);
-        retirados = totalInscritos - efectivos;
-        porcentajeRetirados = Redondeo.redondear_HALFUP(((retirados.doubleValue() / totalInscritos.doubleValue()) * 100.0), 2);
+                List<Inscrito> listaMujeres = listaInscritos.stream().filter(i -> i.getEstudiante().getSexo().equals(Sexo.FEMENINO)).collect(Collectors.toList());
+                mujeres = listaMujeres.size();
+                porcentajeMujeres = Redondeo.redondear_HALFUP(((mujeres.doubleValue() / totalInscritos.doubleValue()) * 100.0), 2);
 
-        reprobados = inscritoFacade.cantidadInscritosReprobados();
-        aprobados = totalInscritos - reprobados;
-        porcentajeAprobados = Redondeo.redondear_HALFUP(((aprobados.doubleValue() / totalInscritos.doubleValue()) * 100.0), 2);
-        porcentajeReprobados = Redondeo.redondear_HALFUP(((reprobados.doubleValue() / totalInscritos.doubleValue()) * 100.0), 2);
+                List<Inscrito> listaRetirados = listaInscritos.stream().filter(i -> i.condicion().equals(Condicion.ABANDONO)).collect(Collectors.toList());
+                retirados = listaRetirados.size();
+                porcentajeRetirados = Redondeo.redondear_HALFUP(((retirados.doubleValue() / totalInscritos.doubleValue()) * 100.0), 2);
 
-        createHorizontalBarModel();
-        createPieModel();
+                efectivos = listaInscritos.size() - listaRetirados.size();
+                porcentajeEfectivos = Redondeo.redondear_HALFUP(((efectivos.doubleValue() / totalInscritos.doubleValue()) * 100.0), 2);
 
-        createDonutModel1();
-        createDonutModel2();
-        createDonutModel3();
+                List<Inscrito> listaAprobados = listaInscritos.stream().filter(i -> i.condicion().equals(Condicion.APROBADO)).collect(Collectors.toList());
+                aprobados = listaAprobados.size();
+                porcentajeAprobados = Redondeo.redondear_HALFUP(((aprobados.doubleValue() / totalInscritos.doubleValue()) * 100.0), 2);
 
-        crearCuadroEstadistico();
+                List<Inscrito> listaReprobados = listaInscritos.stream().filter(i -> i.condicion().equals(Condicion.REPROBADO)).collect(Collectors.toList());
+                reprobados = listaReprobados.size();
+                porcentajeReprobados = Redondeo.redondear_HALFUP(((reprobados.doubleValue() / totalInscritos.doubleValue()) * 100.0), 2);
+
+                createHorizontalBarModel();
+                createPieModel();
+
+                createDonutModel1();
+                createDonutModel2();
+                createDonutModel3();
+
+                crearCuadroEstadistico();
+            } else {
+                varones = 0;
+                porcentajeVarones = 0.0;
+
+                mujeres = 0;
+                porcentajeMujeres = 0.0;
+
+                retirados = 0;
+                porcentajeRetirados = 0.0;
+
+                efectivos = 0;
+                porcentajeEfectivos = 0.0;
+
+                aprobados = 0;
+                porcentajeAprobados = 0.0;
+
+                reprobados = 0;
+                porcentajeReprobados = 0.0;
+
+                createHorizontalBarModel();
+                createPieModel();
+
+                createDonutModel1();
+                createDonutModel2();
+                createDonutModel3();
+
+                crearCuadroEstadistico();
+            }
+        }
+    }
+
+    @Override
+    public List<GestionAcademica> listaGestionesAcademicas() {
+        return gestionAcademicaFacade.listaGestionAcademicaDashboard();
     }
 
     public void createHorizontalBarModel() {
@@ -132,11 +184,14 @@ public class DashboardController extends AbstractController implements Serializa
         HorizontalBarChartDataSet hbarDataSet = new HorizontalBarChartDataSet();
         hbarDataSet.setLabel("Cantidad");
 
-        List<Carrera> carreras = carreraFacade.listaCarreras();
+        List<Carrera> carreras = new ArrayList<>();
+        if (seleccionGestionAcademica != null) {
+            carreras = carreraFacade.listaCarreras(seleccionGestionAcademica.getRegimen());
+        }
 
         List<Number> values = new ArrayList<>();
         for (Carrera carrera : carreras) {
-            Long cantidadInscritos = inscritoFacade.cantidadInscritos(carrera.getId_carrera());
+            Long cantidadInscritos = inscritoFacade.cantidadInscritos(seleccionGestionAcademica.getId_gestionacademica(), carrera.getId_carrera());
             values.add(cantidadInscritos);
         }
         hbarDataSet.setData(values);
@@ -185,11 +240,15 @@ public class DashboardController extends AbstractController implements Serializa
         ChartData data = new ChartData();
 
         DonutChartDataSet dataSet = new DonutChartDataSet();
-        List<Number> values = new ArrayList<>();
 
-        List<Carrera> carreras = carreraFacade.listaCarreras();
+        List<Carrera> carreras = new ArrayList<>();
+        if (seleccionGestionAcademica != null) {
+            carreras = carreraFacade.listaCarreras(seleccionGestionAcademica.getRegimen());
+        }
+
+        List<Number> values = new ArrayList<>();
         for (Carrera carrera : carreras) {
-            Long cantidadInscritos = inscritoFacade.cantidadInscritos(carrera.getId_carrera());
+            Long cantidadInscritos = inscritoFacade.cantidadInscritos(seleccionGestionAcademica.getId_gestionacademica(), carrera.getId_carrera());
             values.add(cantidadInscritos);
         }
 
@@ -289,75 +348,211 @@ public class DashboardController extends AbstractController implements Serializa
     }
 
     public void crearCuadroEstadistico() {
-        nivelesCuadroEstadistico = new ArrayList();
+        if (seleccionGestionAcademica != null) {
+            nivelesCuadroEstadistico = new ArrayList();
+            NivelCuadroEstadistico nivelCuadroEstadistico;
 
-        List<Carrera> carreras = carreraFacade.listaCarreras();
-        for (Carrera carrera : carreras) {
-            Nivel[] niveles = Nivel.values(carrera.getRegimen());
-            for (Nivel nivel : niveles) {
-                NivelCuadroEstadistico nivelCuadroEstadistico = new NivelCuadroEstadistico();
+            Integer totalInscritos = 0;
+            Integer totalInscritosVarones = 0;
+            Integer totalInscritosMujeres = 0;
+            Integer totalRetirados = 0;
+            Integer totalRetiradosVarones = 0;
+            Integer totalRetiradosMujeres = 0;
+            Integer totalEfectivos = 0;
+            Integer totalEfectivosVarones = 0;
+            Integer totalEfectivosMujeres = 0;
+            Integer totalAprobados = 0;
+            Integer totalAprobadosVarones = 0;
+            Integer totalAprobadosMujeres = 0;
+            Integer totalReprobados = 0;
+            Integer totalReprobadosVarones = 0;
+            Integer totalReprobadosMujeres = 0;
 
-                nivelCuadroEstadistico.setCodigo(carrera.getCodigo());
-                nivelCuadroEstadistico.setNombre(carrera.getNombre());
-                nivelCuadroEstadistico.setNivel(nivel);
+            List<Carrera> carreras = carreraFacade.listaCarreras(seleccionGestionAcademica.getRegimen());
+            for (Carrera carrera : carreras) {
 
-                Long numeroParalelos = grupoFacade.cantidadGrupos(carrera.getId_carrera(), nivel);
-                nivelCuadroEstadistico.setNumeroParalelos(numeroParalelos);
+                Integer subtotalInscritos = 0;
+                Integer subtotalInscritosVarones = 0;
+                Integer subtotalInscritosMujeres = 0;
+                Integer subtotalRetirados = 0;
+                Integer subtotalRetiradosVarones = 0;
+                Integer subtotalRetiradosMujeres = 0;
+                Integer subtotalEfectivos = 0;
+                Integer subtotalEfectivosVarones = 0;
+                Integer subtotalEfectivosMujeres = 0;
+                Integer subtotalAprobados = 0;
+                Integer subtotalAprobadosVarones = 0;
+                Integer subtotalAprobadosMujeres = 0;
+                Integer subtotalReprobados = 0;
+                Integer subtotalReprobadosVarones = 0;
+                Integer subtotalReprobadosMujeres = 0;
 
-                Long totalInscritos = inscritoFacade.cantidadInscritos(carrera.getId_carrera(), nivel);
-                nivelCuadroEstadistico.setTotalInscritos(totalInscritos);
+                Nivel[] niveles = Nivel.values(carrera.getRegimen());
+                for (Nivel nivel : niveles) {
 
-                Long varones = inscritoFacade.cantidadInscritos(carrera.getId_carrera(), nivel, Sexo.MASCULINO);
-                nivelCuadroEstadistico.setVarones(varones);
-                Long mujeres = inscritoFacade.cantidadInscritos(carrera.getId_carrera(), nivel, Sexo.FEMENINO);
-                nivelCuadroEstadistico.setMujeres(mujeres);
+                    List<String> paralelos = grupoFacade.listaParalelos(seleccionGestionAcademica.getId_gestionacademica(), carrera.getId_carrera(), nivel);
+                    for (String paralelo : paralelos) {
 
-                Long efectivos = inscritoFacade.cantidadInscritosEfectivos(carrera.getId_carrera(), nivel);
-                nivelCuadroEstadistico.setEfectivos(efectivos);
-                Long retirados = totalInscritos - efectivos;
-                nivelCuadroEstadistico.setRetirados(retirados);
+                        List<Inscrito> inscritos = inscritoFacade.listaInscritos(seleccionGestionAcademica.getId_gestionacademica(), carrera.getId_carrera(), nivel, paralelo);
+                        if (!inscritos.isEmpty()) {
+                            nivelCuadroEstadistico = new NivelCuadroEstadistico();
 
-                Long reprobados = inscritoFacade.cantidadInscritosReprobados(carrera.getId_carrera(), nivel);
-                Long aprobados = totalInscritos - reprobados;
-                nivelCuadroEstadistico.setAprobados(aprobados);
-                nivelCuadroEstadistico.setReprobados(reprobados);
+                            nivelCuadroEstadistico.setCodigo(carrera.getCodigo());
+                            nivelCuadroEstadistico.setNombre(carrera.getNombre());
+                            nivelCuadroEstadistico.setNivel(nivel);
+                            nivelCuadroEstadistico.setParalelo(paralelo);
 
+                            List<Inscrito> inscritosVarones = inscritos.stream().filter(i -> i.getEstudiante().getSexo().equals(Sexo.MASCULINO)).collect(Collectors.toList());
+                            List<Inscrito> inscritosMujeres = inscritos.stream().filter(i -> i.getEstudiante().getSexo().equals(Sexo.FEMENINO)).collect(Collectors.toList());
+                            nivelCuadroEstadistico.setInscritos(inscritos.size());
+                            nivelCuadroEstadistico.setInscritosVarones(inscritosVarones.size());
+                            nivelCuadroEstadistico.setInscritosMujeres(inscritosMujeres.size());
+                            subtotalInscritos += inscritos.size();
+                            totalInscritos += inscritos.size();
+                            subtotalInscritosVarones += inscritosVarones.size();
+                            totalInscritosVarones += inscritosVarones.size();
+                            subtotalInscritosMujeres += inscritosMujeres.size();
+                            totalInscritosMujeres += inscritosMujeres.size();
+
+                            List<Inscrito> retirados = inscritos.stream().filter(i -> i.condicion().equals(Condicion.ABANDONO)).collect(Collectors.toList());
+                            List<Inscrito> retiradosVarones = retirados.stream().filter(i -> i.getEstudiante().getSexo().equals(Sexo.MASCULINO)).collect(Collectors.toList());
+                            List<Inscrito> retiradosMujeres = retirados.stream().filter(i -> i.getEstudiante().getSexo().equals(Sexo.FEMENINO)).collect(Collectors.toList());
+                            nivelCuadroEstadistico.setRetirados(retirados.size());
+                            nivelCuadroEstadistico.setRetiradosVarones(retiradosVarones.size());
+                            nivelCuadroEstadistico.setRetiradosMujeres(retiradosMujeres.size());
+                            subtotalRetirados += retirados.size();
+                            totalRetirados += retirados.size();
+                            subtotalRetiradosVarones += retiradosVarones.size();
+                            totalRetiradosVarones += retiradosVarones.size();
+                            subtotalRetiradosMujeres += retiradosMujeres.size();
+                            totalRetiradosMujeres += retiradosMujeres.size();
+
+                            Integer efectivos = inscritos.size() - retirados.size();
+                            Integer efectivosVarones = inscritosVarones.size() - retiradosVarones.size();
+                            Integer efectivosMujeres = inscritosMujeres.size() - retiradosMujeres.size();
+                            nivelCuadroEstadistico.setEfectivos(efectivos);
+                            nivelCuadroEstadistico.setEfectivosVarones(efectivosVarones);
+                            nivelCuadroEstadistico.setEfectivosMujeres(efectivosMujeres);
+                            subtotalEfectivos += efectivos;
+                            totalEfectivos += efectivos;
+                            subtotalEfectivosVarones += efectivosVarones;
+                            totalEfectivosVarones += efectivosVarones;
+                            subtotalEfectivosMujeres += efectivosMujeres;
+                            totalEfectivosMujeres += efectivosMujeres;
+
+                            List<Inscrito> aprobados = inscritos.stream().filter(i -> i.condicion().equals(Condicion.APROBADO)).collect(Collectors.toList());
+                            List<Inscrito> aprobadosVarones = aprobados.stream().filter(i -> i.getEstudiante().getSexo().equals(Sexo.MASCULINO)).collect(Collectors.toList());
+                            List<Inscrito> aprobadosMujeres = aprobados.stream().filter(i -> i.getEstudiante().getSexo().equals(Sexo.FEMENINO)).collect(Collectors.toList());
+                            nivelCuadroEstadistico.setAprobados(aprobados.size());
+                            nivelCuadroEstadistico.setAprobadosVarones(aprobadosVarones.size());
+                            nivelCuadroEstadistico.setAprobadosMujeres(aprobadosMujeres.size());
+                            subtotalAprobados += aprobados.size();
+                            totalAprobados += aprobados.size();
+                            subtotalAprobadosVarones += aprobadosVarones.size();
+                            totalAprobadosVarones += aprobadosVarones.size();
+                            subtotalAprobadosMujeres += aprobadosMujeres.size();
+                            totalAprobadosMujeres += aprobadosMujeres.size();
+
+                            List<Inscrito> reprobados = inscritos.stream().filter(i -> i.condicion().equals(Condicion.REPROBADO)).collect(Collectors.toList());
+                            List<Inscrito> reprobadosVarones = reprobados.stream().filter(i -> i.getEstudiante().getSexo().equals(Sexo.MASCULINO)).collect(Collectors.toList());
+                            List<Inscrito> reprobadosMujeres = reprobados.stream().filter(i -> i.getEstudiante().getSexo().equals(Sexo.FEMENINO)).collect(Collectors.toList());
+                            nivelCuadroEstadistico.setReprobados(reprobados.size());
+                            nivelCuadroEstadistico.setReprobadosVarones(reprobadosVarones.size());
+                            nivelCuadroEstadistico.setReprobadosMujeres(reprobadosMujeres.size());
+                            subtotalReprobados += reprobados.size();
+                            totalReprobados += reprobados.size();
+                            subtotalReprobadosVarones += reprobadosVarones.size();
+                            totalReprobadosVarones += reprobadosVarones.size();
+                            subtotalReprobadosMujeres += reprobadosMujeres.size();
+                            totalReprobadosMujeres += reprobadosMujeres.size();
+
+                            nivelesCuadroEstadistico.add(nivelCuadroEstadistico);
+                        }
+                    }
+                }
+
+                //subtotales
+                if (subtotalInscritos > 0) {
+                    nivelCuadroEstadistico = new NivelCuadroEstadistico();
+                    nivelCuadroEstadistico.setCodigo("");
+                    nivelCuadroEstadistico.setNombre("");
+                    nivelCuadroEstadistico.setNivel(null);
+                    nivelCuadroEstadistico.setInscritos(subtotalInscritos);
+                    nivelCuadroEstadistico.setInscritosVarones(subtotalInscritosVarones);
+                    nivelCuadroEstadistico.setInscritosMujeres(subtotalInscritosMujeres);
+                    nivelCuadroEstadistico.setRetirados(subtotalRetirados);
+                    nivelCuadroEstadistico.setRetiradosVarones(subtotalRetiradosVarones);
+                    nivelCuadroEstadistico.setRetiradosMujeres(subtotalRetiradosMujeres);
+                    nivelCuadroEstadistico.setEfectivos(subtotalEfectivos);
+                    nivelCuadroEstadistico.setEfectivosVarones(subtotalEfectivosVarones);
+                    nivelCuadroEstadistico.setEfectivosMujeres(subtotalEfectivosMujeres);
+                    nivelCuadroEstadistico.setAprobados(subtotalAprobados);
+                    nivelCuadroEstadistico.setAprobadosVarones(subtotalAprobadosVarones);
+                    nivelCuadroEstadistico.setAprobadosMujeres(subtotalAprobadosMujeres);
+                    nivelCuadroEstadistico.setReprobados(subtotalReprobados);
+                    nivelCuadroEstadistico.setReprobadosVarones(subtotalReprobadosVarones);
+                    nivelCuadroEstadistico.setReprobadosMujeres(subtotalReprobadosMujeres);
+                    nivelesCuadroEstadistico.add(nivelCuadroEstadistico);
+                }
+            }
+            
+            //totales
+            if (totalInscritos > 0) {
+                nivelCuadroEstadistico = new NivelCuadroEstadistico();
+                nivelCuadroEstadistico.setCodigo("");
+                nivelCuadroEstadistico.setNombre("");
+                nivelCuadroEstadistico.setNivel(null);
+                nivelCuadroEstadistico.setInscritos(totalInscritos);
+                nivelCuadroEstadistico.setInscritosVarones(totalInscritosVarones);
+                nivelCuadroEstadistico.setInscritosMujeres(totalInscritosMujeres);
+                nivelCuadroEstadistico.setRetirados(totalRetirados);
+                nivelCuadroEstadistico.setRetiradosVarones(totalRetiradosVarones);
+                nivelCuadroEstadistico.setRetiradosMujeres(totalRetiradosMujeres);
+                nivelCuadroEstadistico.setEfectivos(totalEfectivos);
+                nivelCuadroEstadistico.setEfectivosVarones(totalEfectivosVarones);
+                nivelCuadroEstadistico.setEfectivosMujeres(totalEfectivosMujeres);
+                nivelCuadroEstadistico.setAprobados(totalAprobados);
+                nivelCuadroEstadistico.setAprobadosVarones(totalAprobadosVarones);
+                nivelCuadroEstadistico.setAprobadosMujeres(totalAprobadosMujeres);
+                nivelCuadroEstadistico.setReprobados(totalReprobados);
+                nivelCuadroEstadistico.setReprobadosVarones(totalReprobadosVarones);
+                nivelCuadroEstadistico.setReprobadosMujeres(totalReprobadosMujeres);
                 nivelesCuadroEstadistico.add(nivelCuadroEstadistico);
             }
         }
     }
-    
+
     public void toDashboard() throws IOException {
         reinit();
-        
+
         this.redireccionarViewId("/reportes/dashboard");
     }
 
     /**
      * @return the totalInscritos
      */
-    public Long getTotalInscritos() {
+    public Integer getTotalInscritos() {
         return totalInscritos;
     }
 
     /**
      * @param totalInscritos the totalInscritos to set
      */
-    public void setTotalInscritos(Long totalInscritos) {
+    public void setTotalInscritos(Integer totalInscritos) {
         this.totalInscritos = totalInscritos;
     }
 
     /**
      * @return the varones
      */
-    public Long getVarones() {
+    public Integer getVarones() {
         return varones;
     }
 
     /**
      * @param varones the varones to set
      */
-    public void setVarones(Long varones) {
+    public void setVarones(Integer varones) {
         this.varones = varones;
     }
 
@@ -378,14 +573,14 @@ public class DashboardController extends AbstractController implements Serializa
     /**
      * @return the mujeres
      */
-    public Long getMujeres() {
+    public Integer getMujeres() {
         return mujeres;
     }
 
     /**
      * @param mujeres the mujeres to set
      */
-    public void setMujeres(Long mujeres) {
+    public void setMujeres(Integer mujeres) {
         this.mujeres = mujeres;
     }
 
@@ -448,14 +643,14 @@ public class DashboardController extends AbstractController implements Serializa
     /**
      * @return the efectivos
      */
-    public Long getEfectivos() {
+    public Integer getEfectivos() {
         return efectivos;
     }
 
     /**
      * @param efectivos the efectivos to set
      */
-    public void setEfectivos(Long efectivos) {
+    public void setEfectivos(Integer efectivos) {
         this.efectivos = efectivos;
     }
 
@@ -476,14 +671,14 @@ public class DashboardController extends AbstractController implements Serializa
     /**
      * @return the retirados
      */
-    public Long getRetirados() {
+    public Integer getRetirados() {
         return retirados;
     }
 
     /**
      * @param retirados the retirados to set
      */
-    public void setRetirados(Long retirados) {
+    public void setRetirados(Integer retirados) {
         this.retirados = retirados;
     }
 
@@ -504,14 +699,14 @@ public class DashboardController extends AbstractController implements Serializa
     /**
      * @return the aprobados
      */
-    public Long getAprobados() {
+    public Integer getAprobados() {
         return aprobados;
     }
 
     /**
      * @param aprobados the aprobados to set
      */
-    public void setAprobados(Long aprobados) {
+    public void setAprobados(Integer aprobados) {
         this.aprobados = aprobados;
     }
 
@@ -532,14 +727,14 @@ public class DashboardController extends AbstractController implements Serializa
     /**
      * @return the reprobados
      */
-    public Long getReprobados() {
+    public Integer getReprobados() {
         return reprobados;
     }
 
     /**
      * @param reprobados the reprobados to set
      */
-    public void setReprobados(Long reprobados) {
+    public void setReprobados(Integer reprobados) {
         this.reprobados = reprobados;
     }
 
@@ -597,5 +792,19 @@ public class DashboardController extends AbstractController implements Serializa
      */
     public void setDonutModel3(DonutChartModel donutModel3) {
         this.donutModel3 = donutModel3;
+    }
+
+    /**
+     * @return the seleccionGestionAcademica
+     */
+    public GestionAcademica getSeleccionGestionAcademica() {
+        return seleccionGestionAcademica;
+    }
+
+    /**
+     * @param seleccionGestionAcademica the seleccionGestionAcademica to set
+     */
+    public void setSeleccionGestionAcademica(GestionAcademica seleccionGestionAcademica) {
+        this.seleccionGestionAcademica = seleccionGestionAcademica;
     }
 }