+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0.2] - 2023-01-29
+### Added
+### Changed
+- Dashboard por gestion academica
+### Removed
+### Fixed
+- Se corrigio el Historial Academico
+
 ## [2.0.1] - 2022-12-19
 ### Added
 ### Changed